pyproject.toml
[project]
name = "model-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.20",
    "langchain-google-genai>=2.1.12",
    "langchain-mcp-adapters>=0.1.10",
    "langchain-openai>=0.3.33",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "rich>=14.1.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
]
